Add Discounts
Discounts are added in a super flexible way.

There are application rules, which essentially are used to dictate whether or not a discount is needed based on the entire basket, and pricing rules, which dictate whether to change price or count of the product to affect overall pricing.

We are starting with 3 discounts--bogo for tea, a category wide coffee price reduction after a bulk amount is purchasd, and a flat reduction for strawberries.

Some initial helpers are also added here--
* buy_x_get_y_free which enables any buy X get Y discounts very easily
* static_price which changes the product to be a new price
* reduce_price which changes the product to be a new price based on a percentage of the old price.