Switch to using Decimal until amount rounds out
I had switched the coffee discount to be buy 2 get 1 free, but the wording of the discount makes it apply category-wide (i.e. CF2 should be discounted too).

If that is truly the case, B2G1 isn't correct, it is actually a 2/3 price reduction instead.

Money was rounding inappropriately, so I switched to decimal and cast to money at the end.